requirements.txt
pandas
requests
python-dotenv
plotly
dash
dash-bootstrap-components
numpy
